requirements.txt
aniso8601==7.0.0
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-filter==23.2
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.4
promise==2.3
psycopg2-binary==2.9.6
PyJWT==1.7.1
python-dotenv==1.0.0
requests==2.31.0
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.3
urllib3==1.26.6
